Migrate ContextMenu constants and utils to design system
