+++ design-system/ContextMenu/ContextMenu.utils.ts
@@ -1,4 +1,4 @@
-import { MENU_ITEM_HEIGHT } from "./constants";
+import { MENU_ITEM_HEIGHT } from "./ContextMenu.constants";
 
 export const calculateMenuHeight = (itemLength: number) => {
   return MENU_ITEM_HEIGHT * itemLength;
