--- utils/contextMenu/types.ts
@@ -1,38 +1,0 @@
-export type TransformOriginAnchorPosition =
-  | "top-right"
-  | "top-left"
-  | "top-center"
-  | "bottom-right"
-  | "bottom-left"
-  | "bottom-center";
-
-export type GestureHandlerProps = {
-  children: React.ReactElement | React.ReactElement[];
-};
-
-export type MenuItemProps = {
-  text: string;
-  icon?: string | (() => React.ReactElement);
-  onPress?: (...args: any[]) => void;
-  isTitle?: boolean;
-  isDestructive?: boolean;
-  withSeparator?: boolean;
-};
-
-export type MenuListProps = {
-  items: MenuItemProps[];
-};
-
-export type MenuInternalProps = {
-  items: MenuItemProps[];
-  itemHeight: number;
-  itemWidth: number;
-  itemY: number;
-  itemX: number;
-  anchorPosition: TransformOriginAnchorPosition;
-  menuHeight: number;
-  transformValue: number;
-  actionParams: {
-    [name: string]: (string | number)[];
-  };
-};
